Restructuring of sql_entity_graph
