+++ pgx-macros/src/operators.rs
@@ -1,4 +1,10 @@
-use pgx_utils::{operator_common::*, sql_entity_graph_generators};
+use pgx_utils::{
+    operator_common::*,
+    sql_entity_graph::{
+        postgres_ord::PostgresOrd,
+        postgres_hash::PostgresHash,
+    },
+};
 
 use quote::ToTokens;
 use syn::DeriveInput;
@@ -21,7 +27,7 @@ pub(crate) fn impl_postgres_ord(ast: DeriveInput) -> syn::Result<proc_macro2::To
     stream.extend(ge(&ast.ident));
     stream.extend(cmp(&ast.ident));
 
-    let sql_graph_entity_item = sql_entity_graph_generators::PostgresOrd::from_derive_input(ast)?;
+    let sql_graph_entity_item = PostgresOrd::from_derive_input(ast)?;
     sql_graph_entity_item.to_tokens(&mut stream);
 
     Ok(stream)
@@ -32,7 +38,7 @@ pub(crate) fn impl_postgres_hash(ast: DeriveInput) -> syn::Result<proc_macro2::T
 
     stream.extend(hash(&ast.ident));
 
-    let sql_graph_entity_item = sql_entity_graph_generators::PostgresHash::from_derive_input(ast)?;
+    let sql_graph_entity_item = PostgresHash::from_derive_input(ast)?;
     sql_graph_entity_item.to_tokens(&mut stream);
 
     Ok(stream)
