+++ pgx-macros/src/rewriter.rs
@@ -3,7 +3,11 @@
 
 extern crate proc_macro;
 
-use pgx_utils::{categorize_return_type, CategorizedType};
+use pgx_utils::{
+    categorize_return_type,
+    CategorizedType,
+    sql_entity_graph::pg_extern::PgExtern,
+};
 use proc_macro2::{Ident, Span};
 use quote::{quote, quote_spanned, ToTokens};
 use std::ops::Deref;
@@ -35,7 +39,7 @@ impl PgGuardRewriter {
     pub fn item_fn(
         &self,
         func: ItemFn,
-        entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        entity_submission: Option<&PgExtern>,
         rewrite_args: bool,
         is_raw: bool,
         no_guard: bool,
@@ -53,7 +57,7 @@ impl PgGuardRewriter {
     fn item_fn_with_rewrite(
         &self,
         mut func: ItemFn,
-        entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        entity_submission: Option<&PgExtern>,
         is_raw: bool,
         no_guard: bool,
     ) -> (proc_macro2::TokenStream, bool) {
@@ -189,7 +193,7 @@ impl PgGuardRewriter {
         generics: &Generics,
         func_call: proc_macro2::TokenStream,
         rewritten_return_type: proc_macro2::TokenStream,
-        sql_graph_entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        sql_graph_entity_submission: Option<&PgExtern>,
         no_guard: bool,
     ) -> proc_macro2::TokenStream {
         let guard = if no_guard {
@@ -217,7 +221,7 @@ impl PgGuardRewriter {
 
     fn impl_tuple_udf(
         mut func: ItemFn,
-        entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        entity_submission: Option<&PgExtern>,
     ) -> proc_macro2::TokenStream {
         let func_span = func.span();
         let return_type = func.sig.output;
@@ -256,7 +260,7 @@ impl PgGuardRewriter {
         func_name_wrapper: Ident,
         generics: &Generics,
         func_call: proc_macro2::TokenStream,
-        sql_graph_entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        sql_graph_entity_submission: Option<&PgExtern>,
         optional: bool,
     ) -> proc_macro2::TokenStream {
         let generic_type = proc_macro2::TokenStream::from_str(types.first().unwrap()).unwrap();
@@ -343,7 +347,7 @@ impl PgGuardRewriter {
         func_name_wrapper: Ident,
         generics: &Generics,
         func_call: proc_macro2::TokenStream,
-        entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        entity_submission: Option<&PgExtern>,
         optional: bool,
     ) -> proc_macro2::TokenStream {
         let numtypes = types.len();
@@ -453,7 +457,7 @@ impl PgGuardRewriter {
     fn item_fn_without_rewrite(
         &self,
         mut func: ItemFn,
-        entity_submission: Option<&pgx_utils::sql_entity_graph_generators::PgExtern>,
+        entity_submission: Option<&PgExtern>,
         no_guard: bool,
     ) -> proc_macro2::TokenStream {
         // remember the original visibility and signature classifications as we want
