+++ pgx-utils/Cargo.toml
@@ -28,11 +28,8 @@ toml = "0.5.8"
 unescape = "0.1.0"
 url = "2.2.2"
 eyre = "0.6.6"
-color-eyre = "0.6.0"
 tracing = "0.1.30"
 tracing-error = "0.2.0"
 tracing-subscriber = { version = "0.3.8", features = [ "env-filter" ] }
-libloading = "0.7.3"
-clap = { version = "3.0.14", features = [ "env", "suggestions", "cargo", "derive" ] }
 petgraph = "0.6.0"
 prettyplease = "0.1.4"
