+++ pgx-utils/src/sql_entity_graph/schema/mod.rs
@@ -1,3 +1,5 @@
+pub mod entity;
+
 use proc_macro2::{Span, TokenStream as TokenStream2};
 use quote::{quote, ToTokens, TokenStreamExt};
 use std::hash::{Hash, Hasher};
@@ -63,18 +65,18 @@ impl ToTokens for Schema {
         );
         updated_content.push(syn::parse_quote! {
                 #[no_mangle]
-                pub extern "C" fn  #sql_graph_entity_fn_name() -> pgx::datum::sql_entity_graph::SqlGraphEntity {
+                pub extern "C" fn  #sql_graph_entity_fn_name() -> ::pgx::utils::sql_entity_graph::SqlGraphEntity {
                 extern crate alloc;
                 use alloc::vec::Vec;
                 use alloc::vec;
-                let submission = pgx::datum::sql_entity_graph::SchemaEntity {
+                let submission = pgx::utils::sql_entity_graph::schema::entity::SchemaEntity {
                         module_path: module_path!(),
                         name: stringify!(#ident),
                         file: file!(),
                         line: line!(),
                     };
-                    pgx::datum::sql_entity_graph::SqlGraphEntity::Schema(submission)
-                }
+                ::pgx::utils::sql_entity_graph::SqlGraphEntity::Schema(submission)
+            }
         });
         let _semi = &self.module.semi;
