--- pgx-utils/src/sql_entity_graph_generators/mod.rs
@@ -1,24 +1,0 @@
-pub(crate) mod extension_sql;
-pub(crate) mod pg_aggregate;
-pub(crate) mod pg_extern;
-pub(crate) mod pg_schema;
-pub(crate) mod pgx_attribute;
-pub(crate) mod positioning_ref;
-pub(crate) mod postgres_enum;
-pub(crate) mod postgres_hash;
-pub(crate) mod postgres_ord;
-pub(crate) mod postgres_type;
-pub(crate) mod to_sql;
-
-pub use super::ExternArgs;
-pub use extension_sql::{ExtensionSql, ExtensionSqlFile, SqlDeclared};
-pub use pg_aggregate::PgAggregate;
-pub use pg_extern::{Argument, PgExtern, PgOperator};
-pub use pg_schema::Schema;
-pub use pgx_attribute::{ArgValue, NameValueArg, PgxArg, PgxAttribute};
-pub use positioning_ref::PositioningRef;
-pub use postgres_enum::PostgresEnum;
-pub use postgres_hash::PostgresHash;
-pub use postgres_ord::PostgresOrd;
-pub use postgres_type::PostgresType;
-pub use to_sql::ToSqlConfig;
